requirements.txt
# Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1

# PDF Generation
reportlab==4.0.7
WeasyPrint==60.1
pdfrw==0.4
PyPDF2==3.0.1
pypdf>=3.17.0  # For TREC template form field filling

# Image Handling
Pillow==10.1.0
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
